// frontend/src/api/generated/model/transactionPatch.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Family Finance Tracker
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionPatchCategoryOverride } from './transactionPatchCategoryOverride';
import type { TransactionPatchSubcategory } from './transactionPatchSubcategory';
import type { TransactionPatchOwnerId } from './transactionPatchOwnerId';
import type { TransactionPatchMerchantOverride } from './transactionPatchMerchantOverride';
import type { TransactionPatchTypeOverride } from './transactionPatchTypeOverride';

export interface TransactionPatch {
  category_override?: TransactionPatchCategoryOverride;
  subcategory?: TransactionPatchSubcategory;
  owner_id?: TransactionPatchOwnerId;
  merchant_override?: TransactionPatchMerchantOverride;
  type_override?: TransactionPatchTypeOverride;
}